move store ts test to monorepo root
