+++ packages/config/src/exports/internal.ts
@@ -1,0 +1,1 @@
+// Nothing here yet.
