+++ packages/store/package.json
@@ -74,14 +74,11 @@
     "@latticexyz/gas-report": "workspace:*",
     "@types/debug": "^4.1.7",
     "@types/ejs": "^3.1.1",
-    "@types/mocha": "^9.1.1",
     "@types/node": "^18.15.11",
-    "@viem/anvil": "^0.0.7",
     "ds-test": "https://github.com/dapphub/ds-test.git#e282159d5170298eb2455a6c05280ab5a73a4ef0",
     "forge-std": "https://github.com/foundry-rs/forge-std.git#74cfb77e308dd188d2f58864aaf44963ae6b88b1",
     "solhint": "^3.3.7",
     "tsup": "^6.7.0",
-    "tsx": "^3.12.6",
-    "vitest": "0.34.6"
+    "tsx": "^3.12.6"
   }
 }
