+++ test-setup/global/anvil.ts
@@ -1,11 +1,11 @@
 import { startProxy as startAnvilProxy } from "@viem/anvil";
-import { anvilHost, anvilPort } from "./common";
+import { anvilHost, anvilPort } from "../common";
 import { execa } from "execa";
 
 export default async function globalSetup(): Promise<() => Promise<void>> {
   console.log("building mock game");
   await execa("pnpm", ["run", "build"], {
-    cwd: `${__dirname}/../../../../test/mock-game-contracts`,
+    cwd: `${__dirname}/../../test/mock-game-contracts`,
   });
 
   const shutdownAnvilProxy = await startAnvilProxy({
